+++ docs/changelog.md
@@ -33,6 +33,7 @@ brief notes about them may accumulate here.
 * Windows: Fixed [ToggleFullScreen](config/lua/keyassignment/ToggleFullScreen.md) so that it once again toggles between full screen and normal placement. [#177](https://github.com/wez/wezterm/issues/177)
 * New: [exit_behavior](config/lua/config/exit_behavior.md) config option to keep panes open after the program has completed. [#499](https://github.com/wez/wezterm/issues/499)
 * Closing the configuration error window no longer requires confirmation
+* New: added `--config name=value` options to `wezterm`, `wezterm-gui` and `wezterm-mux-server`
 
 ### 20210203-095643-70a364eb